Updates - More Info & ScanCube
Added three pages on More info - becoming quick start, gives guides on how to use the program, and understanding basic principles of using and solving the rubiks cube (notation etc), also made the instructions on the cube scanner more clear, adding better instructions and clearer text for users to understand. Also added a termiate button, for the scanner, that allows the user to close the scanner at any poin, this was a bug that i found during testing, that the scanner could not be closed once opened.